add transitiontime to setLightState when using bri_inc (dim step)
Motivation: The illuminize zigbee dimmer for some reason does not dim
smoothly when using the "move to level" (0x00) zigbee command with
a transition time. However it does dim smoothly when using the
"step" (0x02) zigbee command instead.

Because deCONZ applies the transition time to the payload correctly,
this can be used as a workaround.